build: :pushpin: Share Inputs
Each flake has its own inputs which in return depend on other flakes. Some flakes even use the same inputs but with different versions, which clutteres the flake.lock file and redundantly stores multiple versions of the same flake. In order to mitigate it this PR introduces a commonly shared flakes in the current flakes and tells the other flakes to reuse the predefined one.